+++ NEWS.md
@@ -1,5 +1,6 @@
 # traits4models 0.2.1
 * New columns 'OriginalReference' and 'OriginalDOI' now possible in harmonized trait data
+* Functions get_trait_data() and get_taxon_data() now return 'OriginalReference' and 'OriginalDOI' if available 
 
 # traits4models 0.2.0
 * New columns 'MininumValue' and 'MaximumValue' added to HarmonizedTraitDefinition